+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- n/a
+- pubtools-exodus-push: Load content using the pushsource library
 
 ## 0.1.0 - 2022-02-24